Make smaller partitions to fit device sizes of 2016.2.x
2016.2.x used smaller partitions, which means that it may have been
installed on devices with very small disks (improbable) or on virtual
machines with virtual disks just fitting the smaller partitions (highly
probable).

Autoupgrading to stock gluon 2018.2.x would fail and break those
devices.

Therefore, for now, go back to smaller partitions.

Note that the old partition sizes were 4MB kernel and 48MB root.  But 4MB
is not enough for current kernels, so the new values are 8MB kernel and
44MB root.