gluon-core: create wireless config from gluon wireless roles section
Delete all wifi-ifaces on reconfigure.
Enable mesh based on the configured wireless role.
Adds logic for client_only_radio detection.

The client_only_radio is extended from the existing draft to theoretically also support 2g multi-band.

If there is an additional radio which does not support the mesh channel, we can configure it on a
different channel. In this case, the other radio on the same band supporting the channel turns off its
client radio and is mesh only.